Cambio en el modelo de datos para la tabla ingredientes a nivel general de la app y sus conexiones con gestor de estados y widgets

+++ lib/data/database/database.dart
@@ -307,7 +307,7 @@ Future<void> venderRecetaTransaction(int recetaId) async {
       // Actualizar la cantidad del ingrediente en la tabla 'Ingredientes'
       await (update(ingredientes)..where((t) => t.id.equals(info.ingrediente.id)))
           .write(IngredientesCompanion(
-             cantidad: Value(nuevoStock),
+             cantidad: Value(nuevoStock.toDouble()),
           ));
     }
